requirements.txt
numpy
matplotlib
python-telegram-bot
